Fix required metadata fields in API docs
